+++ adapters/wagmi-connector/src/bloctoConnector.ts
@@ -5,20 +5,22 @@ import {
   WalletClient,
   ConnectorNotFoundError,
 } from '@wagmi/core';
-import { SwitchChainError, Address, createWalletClient, custom } from 'viem';
+import {
+  SwitchChainError,
+  Address,
+  createWalletClient,
+  custom,
+  numberToHex,
+} from 'viem';
 import type {
   EthereumProviderConfig,
   EthereumProviderInterface as BloctoProvider,
 } from '@blocto/sdk';
 import BloctoSDK from '@blocto/sdk';
-import { providers } from 'ethers';
-import { hexValue } from 'ethers/lib/utils.js';
 import { normalizeChainId } from './util/normalizeChainId';
 
-type BloctoWalletSigner = providers.JsonRpcSigner;
-
 type BloctoOptions = {
-    /**
+  /**
    * Your app’s unique identifier that can be obtained at https://developers.blocto.app,
    * To get advanced features and support with Blocto.
    *
@@ -28,7 +30,7 @@ type BloctoOptions = {
 
   /**
    * @deprecated Use `Web3Modal.defaultChain` instead.
-   * 
+   *
    * The chain ID of the chain to connect to.
    */
   chainId?: number;
@@ -37,13 +39,12 @@ type BloctoOptions = {
    * Custom RPC endpoint.
    */
   rpc?: string;
-}
+};
 
 class BloctoConnector extends Connector<BloctoProvider, BloctoOptions> {
-  readonly id = 'bloctoWallet';
-  readonly name = 'Blocto Wallet';
+  readonly id = 'blocto';
+  readonly name = 'Blocto';
   readonly ready = true;
-
   #provider?: BloctoProvider;
   #onAccountsChangedBind: typeof this.onAccountsChanged;
   #onChainChangedBind: typeof this.onChainChanged;
@@ -138,25 +139,16 @@ class BloctoConnector extends Connector<BloctoProvider, BloctoOptions> {
     return normalizeChainId(chainId);
   }
 
-  async getSigner({
-    chainId,
-  }: { chainId?: number | undefined } = {}): Promise<BloctoWalletSigner> {
-    const [provider, account] = await Promise.all([
-      this.getProvider(),
-      this.getAccount(),
-    ]);
-
-    return new providers.Web3Provider(provider, chainId).getSigner(account);
-  }
-
   async isAuthorized(): Promise<boolean> {
-    const account = await this.getAccount();
-    return Promise.resolve(!!account);
+    const walletName = this.storage?.getItem('wallet');
+    const connected = Boolean(this.storage?.getItem('connected'));
+    const isConnect = walletName === 'blocto' && connected;
+    return Promise.resolve(isConnect);
   }
 
   async switchChain(chainId: number): Promise<Chain> {
     const provider = await this.getProvider();
-    const id = hexValue(chainId);
+    const id = numberToHex(chainId);
     const chain = this.chains.find((x) => x.id === chainId);
     const isBloctoSupportChain =
       provider._blocto.supportNetworkList[`${chainId}`];